Updating example stellar_diam_and_jitter_keywords dictionary
Separated the star-specific information from the distribution-specific information to make it easier for different stars to use the same offset source distribution. This change is in anticipation of adding more stars to the list of available stars.

Also cleaned up a typo and removed a TODO label for a task that has been completed.